back/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic
python-dotenv
openai-agents
python-multipart
motor